Initial V1 newterra w/ Terraform Builder
This is the initial commit of the terraform builder tool, that will build a series of different resources based on your response to questions asked in the command line.  Very helpful for things like creating 5 ec2 instances and 5 s3 buckets very quickly all in a custom vpc/subnet.  And getting you a good usable config running and of the ground.   Terraform-importer was left out of this commit because it is not ready for a beginning user just yet.